Enabling Windows integration tests
(cherry picked from commit 4c8b1fd5a2803e393ad1296692533b7b5c727918)
