+++ integration/image/import_test.go
@@ -11,7 +11,7 @@ import (
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/internal/test/request"
 	"github.com/docker/docker/internal/testutil"
-	"github.com/gotestyourself/gotestyourself/skip"
+	"gotest.tools/skip"
 )
 
 // Ensure we don't regress on CVE-2017-14992.
