+++ integration/internal/requirement/requirement.go
@@ -5,9 +5,6 @@ import (
 	"strings"
 	"testing"
 	"time"
-
-	"github.com/docker/docker/pkg/parsers/kernel"
-	"gotest.tools/icmd"
 )
 
 // HasHubConnectivity checks to see if https://hub.docker.com is
@@ -27,27 +24,4 @@ func HasHubConnectivity(t *testing.T) bool {
 		resp.Body.Close()
 	}
 	return err == nil
-}
-
-func overlayFSSupported() bool {
-	result := icmd.RunCommand("/bin/sh", "-c", "cat /proc/filesystems")
-	if result.Error != nil {
-		return false
-	}
-	return strings.Contains(result.Combined(), "overlay\n")
-}
-
-// Overlay2Supported returns true if the current system supports overlay2 as graphdriver
-func Overlay2Supported(kernelVersion string) bool {
-	if !overlayFSSupported() {
-		return false
-	}
-
-	daemonV, err := kernel.ParseRelease(kernelVersion)
-	if err != nil {
-		return false
-	}
-	requiredV := kernel.VersionInfo{Kernel: 4}
-	return kernel.CompareKernelVersion(*daemonV, requiredV) > -1
-
 }
