+++ integration/plugin/logging/logging_linux_test.go
